requirements.txt
botorch==0.10.0
dash==2.16.1
fairseq==0.12.2
botorch==0.9.5
dash==2.16.1
gauche==0.1.6
gpytorch==1.11
GraKeL==0.1.10
graphein==1.7.5
h5py==3.10.0
matplotlib==3.8.3
mol_ga==0.1.1
molecule_generation==0.4.0
networkx==3.2.1
numpy==1.26.4
ogb==1.3.6
optuna==3.6.1
pandas==2.2.2
plotly==5.21.0
PyTDC==0.4.12
PyYAML==6.0.1
rdkit==2023.9.5
scikit_learn==1.2.2
tensorflow==2.16.1
torch==2.1.2
torch_geometric==2.5.3
tqdm==4.65.0
transformers==4.37.2
